DCiFR Version 2 - DeepFace and FairFace
This is the code for the second version of DCiFR, in which users will have the option to select between facial analysis models based on the results that they desire for their pictures. Users no longer select for the specific attributes that they want to analyze for, but instead upload a folder of photos for analysis of the full capabilities of either model offered.