added progress bar to both analysis pages
Both the DeepFace and FairFace pipelines now have a progress bar imbedded in their analysis loops, scaled to the number of photos uploaded by the user. The progress bar will increase as each photo is analyzed and results are created, and will fill to 100% when all analysis is done. The user can then click to the next page and access their DCiFR results.