results save to DCIFR folder
Using the path module from pathlib, DCIFR is now able to access the path to the user's downloads folder and creates a DCIFR folder to save both Deepface and Fairface CSV results. This folder is only created if it does not exist, and if it does, the code will continue as expected and save to the already existing DCIFR folder.